Parametrized delete org & hub
