+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@alkemio/cli",
-  "version": "0.4.0",
+  "version": "0.4.1",
   "author": "Alkemio Foundation",
   "private": false,
   "license": "EUPL-1.2",
@@ -20,7 +20,7 @@
     "lint:fix": "tsc --noEmit && eslint src/**/*.ts{,x} --fix",
     "nodemon": "nodemon",
     "delete-hub-force": "ts-node-dev src/delete-hub-force.ts",
-    "delete-organizations": "ts-node-dev src/delete-organizations.ts",
+    "delete-organization": "ts-node-dev src/delete-organization.ts",
     "remove-user-from-organizations": "ts-node-dev src/remove-user-from-organizations.ts",
     "authorization-reset-all": "ts-node-dev src/authorization-reset-all.ts",
     "authorization-reset-all-users": "ts-node-dev src/authorization-reset-all-users.ts",
@@ -48,7 +48,7 @@
     "ts-node-dev": "^1.1.7"
   },
   "dependencies": {
-    "@alkemio/client-lib": "^0.14.9",
+    "@alkemio/client-lib": "^0.14.11",
     "dotenv": "^8.2.0",
     "nodemon": "^2.0.6",
     "rimraf": "^3.0.2",
