+++ src/delete-hub-force.ts
@@ -3,10 +3,11 @@ import * as dotenv from 'dotenv';
 import { alkemioClientFactory } from './utils/alkemio-client.factory';
 
 const main = async () => {
-  await deleteOrganizations();
+  const hubID = process.argv[2];
+  await deleteHub(hubID);
 };
 
-export const deleteOrganizations = async () => {
+export const deleteHub = async (hubID: string) => {
   dotenv.config();
   const logger = createLogger();
 
@@ -14,12 +15,10 @@ export const deleteOrganizations = async () => {
   logger.info(`Alkemio server: ${alClient.config.apiEndpointPrivateGraphql}`);
   await alClient.validateConnection();
 
-  const HUB_ID = 'digileefomgevingdemo2';
-
-  const hubInfo = await alClient.hubInfo(HUB_ID);
+  const hubInfo = await alClient.hubInfo(hubID);
   logger.info(`Hub information: ${hubInfo?.nameID}`);
   if (hubInfo?.nameID) {
-    const opportunities = await alClient.opportunities(HUB_ID);
+    const opportunities = await alClient.opportunities(hubID);
     if (opportunities) {
       for (const opportunity of opportunities) {
         logger.info(`==> Opportunity: ${opportunity?.nameID}`);
@@ -30,7 +29,7 @@ export const deleteOrganizations = async () => {
         });
       }
     }
-    const challenges = await alClient.challenges(HUB_ID);
+    const challenges = await alClient.challenges(hubID);
     if (challenges) {
       for (const challenge of challenges) {
         logger.info(`==> Challenge: ${challenge?.nameID}`);
